pyproject.toml
[project]
name = "telbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pytelegrambotapi>=4.23.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "supabase>=2.6.0",
]
